requirements.txt
angr>=9.2.130,<10
angr_utils>=0.5.0,<0.6
capstone>=5.0.0,<6
lark>=1.1,<2
nose>=1.3,<2
numpy>=1.26,<3
Pillow>=10,<13
pwntools>=4.12,<5
rustworkx>=0.14,<1
